Renamed to `useSetup`
